finverse_api/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
pydantic-settings==2.0.3
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.4
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.0.1
alembic==1.12.1
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
# Blockchain sync dependencies
web3==6.9.0
websockets==11.0.3